+++ core/config.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass
+class Config:
+    debug: bool = False
+
+cfg = Config()
