+++ ereignislog.txt
@@ -119,3 +119,4 @@
 [2025-07-27] Laientipps Teil 97 und 98 ergaenzt (Aufgaben 'Weiterfuehrende Laientipps Teil 97 ergaenzen', 'Weiterfuehrende Laientipps Teil 98 ergaenzen' erledigt)
 [2025-07-27] Laientipps Teil 99 und 100 ergaenzt (Aufgaben 'Weiterfuehrende Laientipps Teil 99 ergaenzen', 'Weiterfuehrende Laientipps Teil 100 ergaenzen' erledigt)
 [2025-07-27] Laientipps Teil 101 und 102 ergaenzt (Aufgaben 'Weiterfuehrende Laientipps Teil 101 ergaenzen', 'Weiterfuehrende Laientipps Teil 102 ergaenzen' erledigt)
+[2025-07-27] Laientipp Teil 103 ergaenzt (Aufgabe 'Weiterfuehrende Laientipps Teil 103 ergaenzen' erledigt)
