+++ fortschrittsinfo.md
@@ -1,3 +1,3 @@
 Ist: 100%
 Soll: 100%
-Tipp: 102 erledigt
+Tipp: 103 erledigt
